Fix issues with DWM
Use a custom struct instead of a class for the margins to fix potential weirdness

Attribute the work to the stack overflow post

Rename the class/method since it does more than just add a drop shadow. It also makes the window border/edges be handled by the OS as well.